pyproject.toml
[tool.poetry]
name = "multimodal-rag"
version = "0.1.0"
description = "Multimodal RAG system for image retrieval based on text queries"
authors = ["YongHyun <chgw2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = ">=2.5.0,<2.7.0"
streamlit = "^1.45.1"
numpy = "^2.2.5"
pdf2image = "^1.17.0"
tqdm = "^4.67.1"
colpali-engine = "^0.3.10"
pymupdf = "^1.23.0"
pandas = "^2.2.0"
pillow = "^10.2.0"
pydantic = "^2.6.0"
scikit-learn = "^1.4.0"
fitz = "^0.0.1.dev2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
